requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiogram==3.15.0
arq==0.26.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
pillow==11.1.0
httpx==0.28.1
pydantic-settings==2.7.1
alembic==1.14.1
python-multipart==0.0.20
aiofiles==24.1.0
tenacity==9.0.0
